#89 - Moving easyqueue.async_queue code to easyqueue.queue
