+++ tests/rabbitmq/test_rabbitmq_connections.py
@@ -1,6 +1,6 @@
 import asynctest
 from asynctest import Mock
-from asyncworker.easyqueue.async_queue import AsyncJsonQueue
+from asyncworker.easyqueue.queue import AsyncJsonQueue
 
 from asyncworker.conf import settings
 from asyncworker.rabbitmq.connection import AMQPConnection
